+++ app/routes/user/index.js
@@ -1,11 +1,13 @@
 var router = require("express").Router();
 
+userRouter = require("./user")
 loginRouter = require("./login");
 logoutRouter = require("./logout");
 signupRouter = require("./signup");
 passwordRouter = require("./password");
 searchRouter = require("./search");
 
+router.use("/", userRouter);
 router.use("/login", loginRouter);
 router.use("/logout", logoutRouter);
 router.use("/signup", signupRouter);
